Refactor params and algorithm in separate files
